# docs/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Where vectors come alive - A modern, visual-first vector database",
    "title": "VectrixDB",
    "version": "2.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chunking": {
      "get": {
        "description": "Every saved chunking run, newest first: when, which golden file, which collections it cut, and each technique's count.\n\nKept beside the retrieval runs, as ``chunking/runs/<id>/report.json`` in\nthe same store, and made where the golden questions live, by\n``compare_chunking`` or a function that runs it. The full report of one\nis ``/api/v1/chunking/{run}``, with ``latest`` for the newest.",
        "operationId": "list_chunking_api_v1_chunking_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Chunking Api V1 Chunking Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chunking",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/chunking/{run}": {
      "get": {
        "description": "One chunking run: every build, each technique at its best, the head to head between them, and the pick.\n\n``golden_download`` says whether the caller can download the golden\nquestions the run used, as for an evaluation run.",
        "operationId": "get_chunking_api_v1_chunking__run__get",
        "parameters": [
          {
            "in": "path",
            "name": "run",
            "required": true,
            "schema": {
              "title": "Run",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Chunking Api V1 Chunking  Run  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chunking",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/chunking/{run}/golden": {
      "get": {
        "description": "The golden file a chunking run used, exactly as it was read. A signed-in admin only, as a person.",
        "operationId": "download_chunking_golden_api_v1_chunking__run__golden_get",
        "parameters": [
          {
            "in": "path",
            "name": "run",
            "required": true,
            "schema": {
              "title": "Run",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Chunking Golden",
        "tags": [
          "evaluations"
        ]
      }
    },
...
    "/api/v1/evaluations": {
      "get": {
        "description": "Every saved run, newest first: when, which golden file, which collections it searched, and each setup's top 10 count.\n\n``offset`` skips that many of the newest, for the next page. The full\nreport of one run is ``/api/v1/evaluations/{run}``, with ``latest`` for\nthe newest.",
        "operationId": "list_evaluations_api_v1_evaluations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Evaluations Api V1 Evaluations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evaluations",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/evaluations/{run}": {
      "get": {
        "description": "One run's report: every setup's numbers, the three picks and the rules they were chosen by.\n\n``golden_download`` says whether the caller can download the golden\nquestions this run used: a signed-in admin, when the file was kept.\n``missing_notes`` says in words what ``missing`` holds, each expected\ndocument an index does not hold, the way the command line says it.",
        "operationId": "get_evaluation_api_v1_evaluations__run__get",
        "parameters": [
          {
            "in": "path",
            "name": "run",
            "required": true,
            "schema": {
              "title": "Run",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Evaluation Api V1 Evaluations  Run  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/evaluations/{run}/golden": {
      "get": {
        "description": "The golden file a run used, exactly as it was read: every question, reference answer and expected id.\n\nA signed-in admin only, as a person. The file is the one whose\nfingerprint the run names, so an old run gives its own questions after\nthe file has changed. A run saved before files were kept, or made from\nquestions in memory, has none.",
        "operationId": "download_golden_api_v1_evaluations__run__golden_get",
        "parameters": [
          {
            "in": "path",
            "name": "run",
            "required": true,
            "schema": {
              "title": "Run",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Golden",
        "tags": [
          "evaluations"
        ]
      }
    },
...
  }
}